requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
python-multipart==0.0.6